pyproject.toml
[tool.poetry]
name = "mr_apollo_2n"
version = "0.1.0"
description = "Package for exhaustively and efficiently mapping and analyzing website structures, identifying all nodes."
authors = ["Jose Miguel Resendiz"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"
requests = "^2"
retry = "^0.9.2"
et = "^0.0.2"
tqdm = "^4.66.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^7.4.3"
requests-mock = "^1.11.0"
pytest-mock = "^3.12.0"
jupyter = "^1.0.0"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
